Pre-loads and cache Jenkins CI plugins list
Getting the list of plugins available on Jenkins CI may take quite
a long time (over 10-20 sec). We do need to pre-load that information
at plugin start-up and cache it in memory.

The list of plugins is typically quite small and does not change often.
Having a 24h TTL is more than acceptable.

Change-Id: Iea38f70bdc68f1712a298270af7906a32ec9c9bd