Revamp the UX, introducing the install button
Adds a entry Plugins>Manage to Gerrit Top-Menu if:
- remote plugin management is enabled
- current user has admin permissions

The list of available plugins includes both version + sha1
allowing to match them against the list of loaded plugins.
Plugins that are already installed and loaded are not shown anymore
in the list of the ones available.

Expose the plugin jar URL for installation and leverage that
information from the UX using the HTTP PUT on the Gerrit
/plugin/<name>.jar API

Change-Id: Ie4e6042c0b33bebbd14b515141776bf91e2c4c0b