plugin-manager: Use relative URL to access Gerrit base URL
The plugin-manager plugin was not working properly for setups
with a reverse-proxy with a non-root base path.

Use a relative URL for calculating the base path, so that
any reverse-proxy configuration with a non-root path can
automatically get to the main location correctly.

Rewrite of Change-Id: If382fe16 with a simplified
implementation.

Bug: Issue 13521
Change-Id: I5181098bfb2d4a27eef3b51fc8780cc8f294d1c3